backend/requirements.txt
# Django Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
kombu==5.3.5

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
PyJWT==2.8.0
cryptography==42.0.2

# API & Web
gunicorn==21.2.0
uvicorn==0.27.0
websockets==12.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# AI & ML Frameworks
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
langgraph==0.0.23
openai==1.12.0
anthropic==0.18.1

# Agent Frameworks
smolagents==0.1.0

# API Integrations
google-search-results==2.4.2
stripe==8.0.0
requests==2.31.0
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
colorlog==6.8.2

# Monitoring & Debugging
sentry-sdk==1.40.0
django-debug-toolbar==4.3.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Documentation
drf-spectacular==0.27.1
